requirements.txt
async-timeout==4.0.3
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
python-dotenv==1.0.0
redis==5.0.1
requests==2.31.0
RPi.GPIO==0.7.1
six==1.16.0
spidev==3.6
spotipy==2.23.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
urllib3==2.0.7
Werkzeug==3.0.1
